Make CachedPlayer its own class
