+++ velocity/src/main/java/ch/andre601/advancedserverlist/velocity/events/VelocityEventWrapper.java
@@ -27,8 +27,8 @@ package ch.andre601.advancedserverlist.velocity.events;
 
 import ch.andre601.advancedserverlist.core.interfaces.core.PluginCore;
 import ch.andre601.advancedserverlist.core.interfaces.events.GenericEventWrapper;
+import ch.andre601.advancedserverlist.core.objects.CachedPlayer;
 import ch.andre601.advancedserverlist.core.profiles.players.GenericPlayer;
-import ch.andre601.advancedserverlist.core.profiles.players.PlayerHandler;
 import ch.andre601.advancedserverlist.core.profiles.replacer.placeholders.PlayerPlaceholders;
 import ch.andre601.advancedserverlist.core.profiles.replacer.placeholders.ServerPlaceholders;
 import ch.andre601.advancedserverlist.velocity.VelocityCore;
@@ -140,7 +140,7 @@ public class VelocityEventWrapper implements GenericEventWrapper<Player, Favicon
     }
     
     @Override
-    public GenericPlayer<Player> createPlayer(PlayerHandler.CachedPlayer player, int protocol){
+    public GenericPlayer<Player> createPlayer(CachedPlayer player, int protocol){
         return new VelocityPlayer(player, protocol);
     }
